Keyboard: Rename macropad to treasure/type9 (#4035)
* rename macropad keyboard to type9 under the treasure directory

* fixup the readme

* rename macropad to type9

* moar renames
