+++ keyboards/treasure/type9/info.json
@@ -1,5 +1,5 @@
 {
-  "keyboard_name": "Macropad",
+  "keyboard_name": "Type-9",
   "url": "",
   "maintainer": "qmk",
   "width": 3,
