+++ keyboards/treasure/type9/keymaps/default/readme.md
@@ -1,0 +1,1 @@
+# The default keymap for Type-9
