+++ keyboards/treasure/type9/readme.md
@@ -1,4 +1,4 @@
-# Macropad
+# Type-9
 
 3x3 Macropad.
 
@@ -8,6 +8,6 @@ Hardware Availability: [Treasure](http://macropad.co/)
 
 Make example for this keyboard (after setting up your build environment):
 
-    make macropad:default
+    make treasure/type9:default
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
